# be/swagger.json
{
    "info": {
        "title": "Taxi Booking API",
        "version": "1.0.0",
        "description": "RESTful API description with Swagger"
    },
    "host": "localhost:3000",
    "swagger": "2.0",
    "paths": {
...
        "/vehicle/add": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Add My Vehicle",
                "description": null,
                "operationId": "addVehicle",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addVehicle"
                        }
                    },
                    {
                        "name": "insurance_image",
                        "in": "formData",
                        "type": "file"
                        
                    },{
                        "name": "registation_image",
                        "in": "formData",
                        "type": "file"
                        
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "tags": [
                    "Authentication",
                    "Profile"
                ],
                "summary": "Get logged in user",
                "description": null,
                "operationId": "profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "User not found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authentication",
                    "Update Profile"
                ],
                "summary": "Get logged in user",
                "description": null,
                "operationId": "updateprofile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "User not found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
     },
    "definitions": {
...
            "addVehicle":{
                "type": "object",
                "required": [
                    "brand_make",
                    "brand_model",
                    "car_type",
                    "car_color",
                    "engine_number",
                    "manufacture_year",
                    "seat_capacity",
                    "insurer",
                    "insurance_image",
                    "insurance_validity",
                    "registration_number",
                    "registation_image"
                ],
                "properties": {
                    "brand_make": {
                        "type": "string"
                    },
                    "brand_model": {
                        "type": "string"
                    },
                    "car_type": {
                        "type": "string"
                    },
                    "car_color": {
                        "type": "string"
                    },
                    "engine_number": {
                        "type": "string"
                    },
                    "manufacture_year": {
                        "type": "number"
                    },
                    "seat_capacity": {
                        "type": "number"
                    },
                    "insurer": {
                        "type": "string"
                    },
                    "insurance_image":{
                        "type": "file"
                    },
                    "insurance_validity":{
                        "type": "string",
                        "format": "date"
                    },
                    "registration_number":{
                        "type":"string"
                    },
                    "registation_image":{
                        "type":"file"
                    }

                }
            },
            "UpdateProfile": {
                "type": "object",
                "required": [
                    "firstname",
                    "lastname",
                    "mobile_number",
                    "password",
                    "role_type"
                ],
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile_number": {
                        "type": "string"
                    }
                    
                }
            },
...
    }
}